refactor(sort): change min_idx variable type to size_t in selection_sort function

+++ 2-selection_sort.c
@@ -10,7 +10,7 @@ void selection_sort(int *array, size_t size)
 {
 	size_t i, j;
 	int swap;
-	int min_idx;
+	size_t min_idx;
 
 	if (array == NULL || size < 2)
 		return;
